+++ database/migrations/2025_02_08_102959_create_reservations_table.php
@@ -7,16 +7,20 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up()
     {
-        Schema::create('reservations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('room_id')->constrained('rooms')->onDelete('cascade'); // Explicitly reference 'rooms' table
-            $table->string('customer_name');
-            $table->string('customer_email');
-            $table->date('check_in');
-            $table->date('check_out');
-            $table->integer('num_guests');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('reservations')) {
+            Schema::create('reservations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('hotel_id')->constrained('hotels')->onDelete('cascade'); // Add hotel_id column
+                $table->foreignId('room_id')->constrained('rooms')->onDelete('cascade'); // Explicitly reference 'rooms' table
+                $table->string('customer_name');
+                $table->string('customer_email');
+                $table->string('customer_phone');
+                $table->date('check_in');
+                $table->date('check_out');
+                $table->integer('num_guests');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
